Silence a warning
The return value of `write` is intentionally unused here. Unfortunately
we cannot just use a cast to `void` because `write` is marked
`gnu::warn_unused_result`, not `nodiscard`, and while a cast to `void`
silences the latter, it does not silence the former (according to the
GCC developers, intentionally). Assigning the return value to a variable
does, and making the variable `maybe_unused` silences the unused
variable warning.