Align ancillary message buffers
It’s not clear whether this is actually necessary or not. However, if
one did not do this, it’s not obvious that `CMSG_FIRSTHDR` would be
guaranteed to return a `cmsghdr` pointer with proper alignment. Doing
this makes that much more likely, while also being harmless if
`CMSG_FIRSTHDR` does somehow guarantee proper alignment.