Fix incorrect print format specifier
`uid_t` is not necessarily `int`. Therefore, `%d` may not necessarily
print it properly. To be safer, cast to `long` and use the corresponding
format specifier, `%ld`. This will still fail if `uid_t` is larger than
`long`, or on very large values if `uid_t` is the same size as `long`
but unsigned, but those are much less likely.